[QoL] Fixed crash when passing a ruleset name to 'shield:validate' or 'shield:validate-ctx'. [QoL] Fixed invalid 'set' and 'block' examples in documentation. [QoL] Documented data descriptors used by 'shield:model'. [QoL] Corrected 'shield:ruleset' signature as the ruleset name is mandatory. [QoL] Clarified throw conditions, value trimming and type coercion in rule documentation.

+++ Cast.php
@@ -32,7 +32,7 @@ class Cast extends Rule
                 break;
 
             case 'float':
-                $val = (double)$val;
+                $val = (float)$val;
                 break;
 
             case 'string': case 'str':
